Create home component to nest maps + details
