+++ frontend/src/components/GoogleMaps/google_maps.jsx
@@ -10,30 +10,8 @@ const mapStyles = {
 class GoogleMap extends React.Component{
   constructor(props) {
     super(props);
-    // this.createMarkers = this.createMarkers.bind(this);
   }
 
-  // componentDidMount() {
-  //   this.props.fetchNearbyPlaces();
-  // }
-
-  // createMarkers() {
-  //   this.props.places.map(place => {
-  //     let lat = place.geometry.location.lat;
-  //     let lng = place.geometry.location.lng;
-  //     return (
-  //       <Marker position={{lat: lat, lng: lng}}/>
-  //     )
-  //   })
-  // }
-
-  // findPlaceInfo() {
-  //   this.props.places.map(place => {
-  //     let place_id = place.
-  //   })
-  // }
-
-  
   render() {
     return (
       <div className="google-map">
