+++ frontend/src/components/home/home_container.js
@@ -3,8 +3,7 @@ import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 
 import { fetchNearbyPlaces } from '../../actions/find_places_actions';
-import GoogleMap from './google_maps';
-
+import Home from '../home/home';
 
 const mapStateToProps = state => ({
   places: state.places
@@ -14,4 +13,4 @@ const mapDispatchToProps = dispatch => ({
   fetchNearbyPlaces: () => dispatch(fetchNearbyPlaces())
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(GoogleMap);
+export default connect(mapStateToProps, mapDispatchToProps)(Home);
